styling

+++ app/assets/javascripts/application.js
@@ -69,11 +69,13 @@ app.controller("MainCtrl", ['$scope', '$http', function ($scope, $http) {
               $scope.currentStore.api_token.hex_value +
               '&id=' + receipt.id;
     var receipts = $scope.currentStore.simple_receipts;
-    receipts.splice(receipts.indexOf(receipt), 1);
-    $http.delete(url).
-      success(function(data, status) {
-        console.log(status, data);
-      });
+    if (confirm("Are you sure?")) {
+      receipts.splice(receipts.indexOf(receipt), 1);
+      $http.delete(url).
+        success(function(data, status) {
+          console.log(status, data);
+        });
+    }
   };
 
   // TODO: better to create a custom directive than manipulate DOM
